Finished structure of edits
While not complete yet, the barebones structure of all of the modifications have been made. I've modified some of the values in residue_constants in order to reflect the different chemical proportions of RNA molecules vs proteins. Substitution isn't complete yet, and a search function for the structure template files needs to be done, but other than those and bug fixes, I believe that this is almost done.